<1,2> feat(portal-base): add experimental ESM build
This is an experimental build that lets you deploy projects, written in
ESM, without the need for any transformation, to Liferay Portal.

It leverages browser modules support for module loading (instead of the AMD
loader).

This is far from being finished, but we will keep updating this until it is
functional or proved unfeasible.