<b> feat(portal-base): support webpack: imports in importmap.json
These are like cjs: imports but instead of pointing to the exact file, they
point to a label that is resolved with the npm algorithm.

We may remove the cjs: imports in the future and leave just webpack: ones.

Additionally, we must decide if webpack: is a good name or if we prefer a
more abstract one. This depends on wether we want to support, for example,
rollup (in addition to webpack) in the future or simply provide an agnostic
way to bundle+bridge CommonJS packages (and we decide internally how to
handle it).

We could also have both things. For example we could have webpack: or
rollup: to force a bundler to be used, then have another agnostic scheme
like cjs: to let the JS Toolkit which bundler to use.