<c> feat(portal-base): add custom-element-webpack-based build
This build bundles the whole project with webpack. This implies that, for
now, we won't be using `react` or `react-dom` from portal, nor will the
developer be able to define any import map.

In the future we may change this but we don't know if it will be necessary
since this type of build should be more like a last resort in case the
others won't work.