<d> feat: support custom-element build
We make a standard webpack build for custom elements and simply declare
packages provided by the portal as `externals`.

We need to support dev and production builds for webpack.

We also still need to find some way to let the user extend webpack and
babel's configuration.

Finally, we should provide a `yarn serve` script to spawn a webpack dev
server for live development.